Reorganize tests for message creators
